Release 11.2.0

+++ src/Operation/Collecting/ForkMatch.php
@@ -7,19 +7,31 @@ use FiiSoft\Jackdaw\Internal\Signal;
 use FiiSoft\Jackdaw\Operation\Collecting\Fork\BaseFork;
 use FiiSoft\Jackdaw\Operation\Collecting\Fork\ForkHandler;
 use FiiSoft\Jackdaw\Operation\Collecting\Fork\ForkHandlerFactory;
+use FiiSoft\Jackdaw\Operation\Exception\OperationExceptionFactory;
 
-final class Fork extends BaseFork
+final class ForkMatch extends BaseFork
 {
-    private ForkHandler $prototype;
+    private ?ForkHandler $prototype = null;
     
     /**
      * @param DiscriminatorReady|callable|array<string|int> $discriminator
+     * @param array<string|int, ForkReady> $handlers
      */
-    public function __construct($discriminator, ForkReady $prototype)
+    public function __construct($discriminator, array $handlers, ?ForkReady $prototype = null)
     {
         parent::__construct($discriminator);
         
-        $this->prototype = ForkHandlerFactory::adaptPrototype($prototype);
+        if (empty($handlers)) {
+            throw OperationExceptionFactory::forkMatchHandlersCannotBeEmpty();
+        }
+        
+        foreach ($handlers as $key => $handler) {
+            $this->handlers[$key] = ForkHandlerFactory::adaptHandler($handler);
+        }
+        
+        if ($prototype !== null) {
+            $this->prototype = ForkHandlerFactory::adaptPrototype($prototype);
+        }
     }
     
     public function handle(Signal $signal): void
@@ -28,9 +40,11 @@ final class Fork extends BaseFork
         
         if (isset($this->handlers[$classifier])) {
             $handler = $this->handlers[$classifier];
-        } else {
+        } elseif ($this->prototype !== null) {
             $handler = $this->prototype->create();
             $this->handlers[$classifier] = $handler;
+        } else {
+            throw OperationExceptionFactory::handlerIsNotDefined($classifier);
         }
         
         $handler->accept($signal->item->value, $signal->item->key);
@@ -46,9 +60,11 @@ final class Fork extends BaseFork
             
             if (isset($this->handlers[$classifier])) {
                 $handler = $this->handlers[$classifier];
-            } else {
+            } elseif ($this->prototype !== null) {
                 $handler = $this->prototype->create();
                 $this->handlers[$classifier] = $handler;
+            } else {
+                throw OperationExceptionFactory::handlerIsNotDefined($classifier);
             }
             
             $handler->accept($value, $key);
@@ -62,7 +78,10 @@ final class Fork extends BaseFork
         if (!$this->isDestroying) {
             parent::destroy();
             
-            $this->prototype->destroy();
+            if ($this->prototype !== null) {
+                $this->prototype->destroy();
+                $this->prototype = null;
+            }
         }
     }
 }
